refactor(Pagination): refactor pagination to a different service
